requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.22
sqlalchemy==2.0.31
alembic==1.13.2
httpx==0.27.0
bcrypt==3.2.2
redis==4.6.0
prometheus-client==0.16.0
tenacity==9.1.2
pytest==7.4.0
pytest-asyncio==0.22.0
aiohttp==3.8.5
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
yfinance==0.2.41  # optional for Yahoo Finance data
asyncpg==0.29.0
psycopg2-binary==2.9.9
requests==2.32.3
aiofiles==24.1.0
apscheduler==3.10.4
loguru==0.7.2
websockets==12.0
tqdm==4.66.4
beautifulsoup4==4.12.3
lxml==5.2.1
newsapi-python==0.2.7
praw==7.7.1
tradingview-ta==3.3.0
openai==1.44.0
joblib==1.5.2
schedule==1.2.1
scikit-learn==1.7.2
PyJWT==2.8.0
cryptography==40.0.2




# 🔑 Auth
python-jose[cryptography]==3.5.0
python-multipart==0.0.6

# 📦 Trading & exchanges
python-binance  # optional but recommended for Binance live execution
ccxt            # optional alternative for exchange integration
       

# 💳 Payments
stripe
paypalrestsdk

# ⚙️ Monitoring / config
sentry-sdk
pydantic-settings

pydantic[email]==2.11.8


fastapi-limiter
